Moves Compiler into a dedicated module
